Move Ring conformance tests to TestExt
