+++ test/generic/MPoly-test.jl
@@ -1697,17 +1697,17 @@ end
 
 @testset "Generic.MPoly.Ring_interface" begin
   S, = polynomial_ring(QQ, 0)
-  test_Ring_interface_recursive(S)
+  ConformanceTests.test_Ring_interface_recursive(S)
 
   S, = polynomial_ring(QQ, 1)
-  test_Ring_interface_recursive(S)
+  ConformanceTests.test_Ring_interface_recursive(S)
 
   S, = polynomial_ring(ZZ, 2)
-  test_Ring_interface_recursive(S)
+  ConformanceTests.test_Ring_interface_recursive(S)
 
   R, = QQ[:x]
   S, = polynomial_ring(R, :z => 1:3)
-  test_Ring_interface(S) # _recursive needs too many ressources
+  ConformanceTests.test_Ring_interface(S) # _recursive needs too many ressources
 end
 
 @testset "Generic.MPoly.zero_rings" begin
@@ -1715,7 +1715,7 @@ end
   S, = polynomial_ring(R, 2)
   @test is_zero(gen(S, 1)) && is_one(gen(S, 1))
   @test is_zero(one(S))
-  test_Ring_interface_recursive(S)
+  ConformanceTests.test_Ring_interface_recursive(S)
 end
 
 # -------------------------------------------------------
