+++ test/generic/RationalFunctionField-test.jl
@@ -4,7 +4,7 @@ end
 
 @testset "Generic.FunctionField.conformance" begin
    S, x = rational_function_field(QQ, "x")
-   test_Ring_interface(S)
+   ConformanceTests.test_Ring_interface(S)
 end
 
 @testset "Generic.RationalFunctionField.constructors" begin
