+++ test/generic/RelSeries-test.jl
@@ -18,7 +18,7 @@
 
 @testset "Generic.RelSeries.conformance" begin
    R, x = power_series_ring(ZZ, 10, "x")
-   test_Ring_interface(R)
+   ConformanceTests.test_Ring_interface(R)
 end
 
 @testset "Generic.RelSeries.types" begin
