+++ test/julia/Integers-test.jl
@@ -1,6 +1,6 @@
 @testset "Julia.Integers.conformance_tests" begin
-   test_Ring_interface_recursive(ZZ)
-   test_EuclideanRing_interface(ZZ)
+   ConformanceTests.test_Ring_interface_recursive(ZZ)
+   ConformanceTests.test_EuclideanRing_interface(ZZ)
 end
 
 @testset "Julia.Integers.manipulation" begin
